SourceAnalysis.java
This is a fairly straightforward piece of software that I submitted for a modern world history class final project. The rubric encouraged students to submit something "compelling and effective," and must "[relate] to or [portray] a topic or theme from world history." I was aiming for simplicity here, and because I would be the only one to actually run the program, I did not concern myself with implementing input filtering or creating a GUI. 
The program reads .txt files from a local directory and outputs information on the file depending on what it is that the user requests.